hw03/requirements.txt
face-recognition==1.3.0
streamlit>=1.20.0
Pillow>=9.0.0
numpy>=1.21.0